backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.4
requests>=2.31
Pillow>=10.0
python-dotenv>=1.0
transliterate

